reworked variable types

+++ ROS Nodes/left_wheels.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int16
+from std_msgs.msg import Int32
 
 def callback(data):
     rospy.loginfo(data.data)
@@ -8,7 +8,7 @@ def callback(data):
 def left_wheels():
 
     rospy.init_node('left_wheels')
-    rospy.Subscriber('left_wheels', Int16, callback)
+    rospy.Subscriber('left_wheels', Int32, callback)
     rospy.spin()
 
 if __name__ == '__main__':
